requirements.txt
fastapi
uvicorn
pydantic
djangorestframework
python-dotenv
pywin32
unidecode
Pillow
